added ability to set 16 bit ints

+++ lib/mlkit.dart
@@ -257,7 +257,7 @@ class FirebaseModelInterpreter {
       {String remoteModelName,
       String localModelName,
       FirebaseModelInputOutputOptions inputOutputOptions,
-      Uint8List inputBytes}) async {
+      Uint16List inputBytes}) async {
     assert(remoteModelName != null || localModelName != null);
     try {
       dynamic results =
